pyproject.toml
[project]
name = "streamlit-yandex-stt"
version = "0.0.1"
description = "Yandex speech to text showcase"
requires-python = ">=3.13"
dependencies = [
    "streamlit==1.52.2",
    "yandex-speechkit==1.5.0",
    "python-dotenv==1.2.1",
    "audioop-lts==0.2.2",
    "pydantic-settings>=2.12.0",
]
